% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakID.r
\name{PeakID}
\alias{PeakID}
\title{Function that identifies peaks in a dataset}
\usage{
PeakID(x, y, w = 1, ...)
}
\arguments{
\item{x}{Vector of \code{x} values of input data}

\item{y}{Vector of \code{y} values of input data}

\item{w}{Window size for smoothing data}
}
\value{
A vector listing the standard deviations of propagated errors
propagated on all \code{X} values.
}
\description{
Developed by William A. Huber
}
\examples{
# Create dummy periodic data
x <- seq(1, 100, 1)
y <- sin((2 * pi * (seq(1, 100, 1) - 8 + 20 / 4)) / 20)
# Run peakID function
result <- PeakID(x, y, w = 20)
}
\references{
package dependencies: zoo 1.8.7
}
\seealso{
https://rpubs.com/mengxu/peak_detection
}
