% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T_from_d18O_d18Ow.R
\name{T_from_d18O_d18Ow}
\alias{T_from_d18O_d18Ow}
\title{Function BLATANTLY STOLEN FROM SHELLCHRON PACKAGE to convert d18O data to SST}
\usage{
T_from_d18O_d18Ow(d18Oc, d18Ow = 0, transfer_function = "KimONeil97")
}
\arguments{
\item{d18Oc}{Matrix with a distance column (values in any unit) and an d18Oc column
(values in permille VPDB)}

\item{d18Ow}{Either a single value (constant d18Ow) or a vector of length
equal to the number of d18O values
Defaults to constant d18Ow of 0 permille VSMOW
(the modern mean ocean value)}

\item{transfer_function}{String containing the name of the transfer function
(for example: \code{"KimONeil97"} or \code{"GrossmanKu86"}). Defaults to
Kim and O'Neil (1997).}
}
\value{
A vector containing SST values for each d18Os value in \code{"d18Oc"}
}
\description{
Takes a matrix of d18Os data (in permille VPDB) against distance measures (in any unit), information
about the d18O value (in permille VSMOW) of the water and how it changes
from one value to another, and the transfer function used for of the record (e.g.
Kim and O'Neil, 1997 or Grossman and Ku, 1986). Converts the d18O data to SST
data (in degrees Celsius) using the supplied empirical transfer function.
}
\examples{
# Create dummy d18Os data
dist <- seq(1, 40, 1) #distance
val <- sin((0.5 * pi * (dist)) / 5)+1 # d18Os
d18O <- cbind(dist, val)
# Run SST model function
SST <- SST_model(d18O, 0, "KimONeil97")
}
\references{
Grossman, E.L., Ku, T., Oxygen and carbon isotope fractionation in biogenic
aragonite: temperature effects, \emph{Chemical Geology} \strong{1986}, \emph{59.1}, 59-74.
\doi{10.1016/0168-9622(86)90057-6}
Kim, S., O'Niel, J.R., Equilibrium and nonequilibrium oxygen
isotope effects in synthetic carbonates, \emph{Geochimica et Cosmochimica Acta}
\strong{1997}, \emph{61.16}, 3461-3475.
\doi{10.1016/S0016-7037(97)00169-5}
Dettman, D.L., Reische, A.K., Lohmann, K.C., Controls on the stable isotope
composition of seasonal growth bands in aragonitic fresh-water bivalves
(Unionidae), \emph{Geochimica et Cosmochimica Acta} \strong{1999}, \emph{63.7-8}, 1049-1057.
\doi{10.1016/S0016-7037(99)00020-4}
Brand, W.A., Coplen, T.B., Vogl, J., Rosner, M., Prohaska, T., Assessment of
international reference materials for isotope-ratio analysis (IUPAC Technical
Report), \emph{Pure and Applied Chemistry} \strong{2014}, \emph{86.3}, 425-467.
\doi{10.1515/pac-2013-1023}
}
